requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
numpy